Split `DisabledReason` enum from `Common.h`
