+++ OPHD/Common.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "EnumConnectorDir.h"
+#include "EnumDisabledReason.h"
 #include "EnumIdleReason.h"
 #include "EnumMoraleIndex.h"
 #include "EnumProductType.h"
@@ -60,19 +61,6 @@ enum class MineProductionRate
 	Low,
 	Medium,
 	High,
-};
-
-
-enum class DisabledReason
-{
-	None, /**< Not Disabled, default reason. */
-
-	Chap, /**< Requires atmosphere, no atmosphere available. */
-	Disconnected, /**< Not connected to Command Center */
-	Energy, /**< Not enough Energy to operate. */
-	Population, /**< Insufficient workers or scientists (or both) */
-	RefinedResources, /**< Insufficient mined and refined resources */
-	StructuralIntegrity /**< Structural integrity out of operating tolerances (damaged structure) */
 };
