+++ OPHD/ophd.vcxproj.filters
@@ -329,6 +329,9 @@
     <ClInclude Include="EnumDirection.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="EnumDisabledReason.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="EnumIdleReason.h">
       <Filter>Header Files</Filter>
     </ClInclude>
